requirements.txt
absl-py==1.4.0
array_record==0.5.1
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.3.0
cycler==0.12.1
dacite==1.8.1
dask==2023.3.1
dataclass_array==1.5.1
decorator==5.2.1
dm-tree==0.1.8
docstring_parser==0.16
einops==0.8.1
einsum==0.3.0
etils==1.5.2
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gast==0.4.0
google-auth==2.16.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
immutabledict==2.2.0
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
joblib==1.4.2
keras==2.12.0
kiwisolver==1.4.7
lark==1.2.2
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.6.1
ml_dtypes==0.5.1
networkx==3.2.1
numpy==1.23.0
oauthlib==3.2.2
opencv-python==4.11.0.86
OpenEXR==1.3.9
opt_einsum==3.4.0
packaging==24.2
pandas==1.5.3
partd==1.4.2
Pillow==9.2.0
plotly==5.13.1
promise==2.3
protobuf==4.21.12
psutil==7.0.0
pyarrow==10.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.6.0
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
simple-parsing==0.1.7
six==1.17.0
tenacity==9.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-addons==0.23.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.12.0
tensorflow-graphics==2021.12.3
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.17.0
tensorflow-probability==0.19.0
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2024.8.30
toml==0.10.2
toolz==1.0.0
tqdm==4.67.1
trimesh==4.6.6
typeguard==2.13.3
typing_extensions==4.13.1
urllib3==2.3.0
visu3d==1.5.1
waymo-open-dataset-tf-2-12-0==1.6.5
Werkzeug==3.1.3
wrapt==1.14.1
zipp==3.21.0
